requirements.txt
fastapi
uvicorn
psycopg2-binary
anthropic
sentence-transformers
numpy
python-dotenv